requirements.txt
altgraph==0.17.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
ConfigArgParse==1.7.1
dotenv==0.9.9
exceptiongroup==1.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
future==0.18.2
gevent==25.9.1
geventhttpclient==2.3.4
greenlet==3.2.4
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.34.0
macholib==1.15.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
packaging==25.0
pluggy==1.6.0
psutil==7.1.1
psycopg2-binary==2.9.10
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dotenv==1.1.1
pyzmq==27.1.0
requests==2.32.5
rich==14.2.0
sentry-sdk==2.43.0
six==1.15.0
SQLAlchemy==2.0.43
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
zope.event==6.0
zope.interface==8.0.1
